--- eq-author/src/components/Comments_v2/graphql/commentSubscription.graphql
@@ -1,5 +1,0 @@
-subscription CommentsUpdated($id: ID!) {
-  commentsUpdated(id: $id) {
-    id
-  }
-}
