--- eq-author/src/components/Comments_v2/graphql/commentsQuery.graphql
@@ -1,28 +1,0 @@
-query Comments($componentId: ID!) {
-  comments(id: $componentId) {
-    id
-    commentText
-    user {
-      id
-      name
-      picture
-      email
-      displayName
-    }
-    createdTime
-    editedTime
-    replies {
-      id
-      commentText
-      createdTime
-      editedTime
-      user {
-        id
-        name
-        picture
-        email
-        displayName
-      }
-    }
-  }
-}
