--- eq-author/src/components/Comments_v2/graphql/createNewComment.graphql
@@ -1,28 +1,0 @@
-mutation createComment($input: CreateCommentInput!) {
-  createComment(input: $input) {
-    id
-    commentText
-    createdTime
-    editedTime
-    user {
-      id
-      name
-      picture
-      email
-      displayName
-    }
-    replies {
-      id
-      commentText
-      createdTime
-      editedTime
-      user {
-        id
-        name
-        picture
-        email
-        displayName
-      }
-    }
-  }
-}
