--- eq-author/src/components/Comments_v2/graphql/createNewReply.graphql
@@ -1,16 +1,0 @@
-mutation CreateReply($input: CreateReplyInput!) {
-  createReply(input: $input) {
-    id
-    commentText
-    createdTime
-    editedTime
-    user {
-      id
-      name
-      picture
-      email
-      displayName
-    }
-    parentCommentId
-  }
-}
