--- eq-author/src/components/Comments_v2/graphql/deleteComment.graphql
@@ -1,28 +1,0 @@
-mutation deleteComment($input: DeleteCommentInput!) {
-  deleteComment(input: $input) {
-    id
-    commentText
-    user {
-      id
-      name
-      picture
-      email
-      displayName
-    }
-    createdTime
-    editedTime
-    replies {
-      id
-      commentText
-      createdTime
-      editedTime
-      user {
-        id
-        name
-        picture
-        email
-        displayName
-      }
-    }
-  }
-}
