--- eq-author/src/components/Comments_v2/graphql/deleteReply.graphql
@@ -1,16 +1,0 @@
-mutation DeleteReply($input: DeleteReplyInput!) {
-  deleteReply(input: $input) {
-    id
-    commentText
-    createdTime
-    editedTime
-    user {
-      id
-      name
-      picture
-      email
-      displayName
-    }
-    parentCommentId
-  }
-}
