--- eq-author/src/components/Comments_v2/graphql/updateComment.graphql
@@ -1,7 +1,0 @@
-mutation updateComment($input: UpdateCommentInput!) {
-  updateComment(input: $input) {
-    id
-    commentText
-    editedTime
-  }
-}
