--- eq-author/src/components/Comments_v2/graphql/updateReply.graphql
@@ -1,7 +1,0 @@
-mutation UpdateReply($input: UpdateReplyInput!) {
-  updateReply(input: $input) {
-    id
-    commentText
-    editedTime
-  }
-}
