+++ eq-author/src/storybook/Components/Comment.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Canvas } from "@storybook/addon-docs/blocks";
 import { ArgsTable } from "@storybook/addon-docs/blocks";
-import Comment from "components/Comments_v2/Comment";
+import Comment from "components/CommentsPanel/Comment";
 
 <Meta title="Components/Comment" component={Comment} />
